+++ zephyr-mcumgr-cli/src/groups/image.rs
@@ -1,3 +1,5 @@
+use indicatif::MultiProgress;
+
 use crate::{
     args::CommonArgs, client::Client, errors::CliError, file_read_write::read_input_file,
     formatting::structured_print, groups::parse_sha256, progress::with_progress_bar,
@@ -30,7 +32,12 @@ pub enum ImageCommand {
     SlotInfo,
 }
 
-pub fn run(client: &Client, args: CommonArgs, command: ImageCommand) -> Result<(), CliError> {
+pub fn run(
+    client: &Client,
+    multiprogress: &MultiProgress,
+    args: CommonArgs,
+    command: ImageCommand,
+) -> Result<(), CliError> {
     let client = client.get()?;
     match command {
         ImageCommand::GetState => {
@@ -64,9 +71,12 @@ pub fn run(client: &Client, args: CommonArgs, command: ImageCommand) -> Result<(
         } => {
             let (data, source_filename) = read_input_file(&image_file)?;
 
-            with_progress_bar(!args.quiet, source_filename.as_deref(), |progress| {
-                client.image_upload(&data, image_id, checksum, upgrade_only, progress)
-            })?;
+            with_progress_bar(
+                multiprogress,
+                !args.quiet,
+                source_filename.as_deref(),
+                |progress| client.image_upload(&data, image_id, checksum, upgrade_only, progress),
+            )?;
         }
         ImageCommand::Erase { slot } => client.image_erase(slot)?,
         ImageCommand::SlotInfo => {