+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add CLI commands:
   - `image`
     - `upload`
+- Refactor cli color output, remove `termcolor` dependency
 
 ## [0.6.2] - 2026-01-13